packages/client/pyproject.toml
[project]
name = "handler-client"
version = "0.1.0"
description = "Shared A2A client utilities for Handler"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.2",
    "a2a-sdk>=0.2.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
